check against domain.key filezize, improved restart procedure

+++ config/letsencrypt/letsrenew.sh
@@ -10,16 +10,28 @@ python /config/letsencrypt/acme_tiny.py --account-key /config/letsencrypt/accoun
 # Removing firewall rule added earlier
 iptables -D $CHAIN 1
 
-# Copying files to lighttpd directory on success
-if [ -s "/config/letsencrypt/domain.key" ]
+# Copying files to lighttpd directory on success only (file domain.key is not empty)
+if [ -f "/config/letsencrypt/domain.key" ] && [ ! $(stat -c %s /config/letsencrypt/domain.key) -eq 0 ]
   then
   cat /config/letsencrypt/signed.crt | tee /etc/lighttpd/server.pem
   cat /config/letsencrypt/domain.key | tee -a /etc/lighttpd/server.pem
 
   # Restarting original lighttpd webserver for EdgeOS
-  ps aux | grep lighttpd.conf | awk '{print $2;}' | head -n1 | xargs kill
-  /usr/sbin/lighttpd -f /etc/lighttpd/lighttpd.conf
+  sudo /sbin/start-stop-daemon --stop --pidfile /var/run/lighttpd.pid
+  if [ -f "/var/run/lighttpd.pid" ]; then
+    rm /var/run/lighttpd.pid
+  fi
+  sudo /sbin/start-stop-daemon --start --quiet \
+        --pidfile /var/run/lighttpd.pid \
+        --exec /usr/sbin/lighttpd -- -f /etc/lighttpd/lighttpd.conf
+  
   # Restart custom lighttpd webserver for custom scripts
-  ps aux | grep lighttpd_custom.conf | awk '{print $2;}' | head -n1 | xargs kill
-  /usr/sbin/lighttpd -f /config/custom/lighttpd/lighttpd_custom.conf
+  sudo /sbin/start-stop-daemon --stop --pidfile /var/run/lighttpd_custom.pid
+  if [ -f "/var/run/lighttpd_custom.pid" ]; then
+    rm /var/run/lighttpd_custom.pid
+  fi
+  sudo /sbin/start-stop-daemon --start --quiet \
+        --pidfile /var/run/lighttpd_custom.pid \
+        --exec /usr/sbin/lighttpd -- -f /config/custom/lighttpd/lighttpd_custom.conf
+  
 fi
